Move system tests and update CI workflow
