+++ tdms/requirements.txt
@@ -1,0 +1,3 @@
+numpy
+pytest
+h5py
